+++ app/schemas/gameenums.py
@@ -2814,6 +2814,7 @@ EVENT_TYPE_NAME: dict[int, NiceEventType] = {
     22: NiceEventType.warBoard,
     23: NiceEventType.combineCosutumeItem,
     24: NiceEventType.treasureBox,
+    21: NiceEventType.comebackCampaign,
 }
 
 
@@ -3160,15 +3161,9 @@ class NiceAiCond(str, Enum):
     checkOpponentHpHigher = "checkOpponentHpHigher"
     checkOpponentHpLower = "checkOpponentHpLower"
     checkTargetPosition = "checkTargetPosition"
-    checkSelfBuffActiveAndPassiveIndividuality = (
-        "checkSelfBuffActiveAndPassiveIndividuality"
-    )
-    checkPtBuffActiveAndPassiveIndividuality = (
-        "checkPtBuffActiveAndPassiveIndividuality"
-    )
-    checkOpponentBuffActiveAndPassiveIndividuality = (
-        "checkOpponentBuffActiveAndPassiveIndividuality"
-    )
+    checkSelfBuffActiveAndPassiveIndividuality = "checkSelfBuffActiveAndPassiveIndividuality"
+    checkPtBuffActiveAndPassiveIndividuality = "checkPtBuffActiveAndPassiveIndividuality"
+    checkOpponentBuffActiveAndPassiveIndividuality = "checkOpponentBuffActiveAndPassiveIndividuality"
     checkPtAllBuff = "checkPtAllBuff"
     checkOpponentAllBuff = "checkOpponentAllBuff"
     checkPtAllBuffIndividuality = "checkPtAllBuffIndividuality"
@@ -3180,23 +3175,17 @@ class NiceAiCond(str, Enum):
     countHigherBuffIndividualitySumPt = "countHigherBuffIndividualitySumPt"
     countHigherBuffIndividualitySumPtAll = "countHigherBuffIndividualitySumPtAll"
     countHigherBuffIndividualitySumOpponent = "countHigherBuffIndividualitySumOpponent"
-    countHigherBuffIndividualitySumOpponentAll = (
-        "countHigherBuffIndividualitySumOpponentAll"
-    )
+    countHigherBuffIndividualitySumOpponentAll = "countHigherBuffIndividualitySumOpponentAll"
     countHigherBuffIndividualitySumSelf = "countHigherBuffIndividualitySumSelf"
     countLowerBuffIndividualitySumPt = "countLowerBuffIndividualitySumPt"
     countLowerBuffIndividualitySumPtAll = "countLowerBuffIndividualitySumPtAll"
     countLowerBuffIndividualitySumOpponent = "countLowerBuffIndividualitySumOpponent"
-    countLowerBuffIndividualitySumOpponentAll = (
-        "countLowerBuffIndividualitySumOpponentAll"
-    )
+    countLowerBuffIndividualitySumOpponentAll = "countLowerBuffIndividualitySumOpponentAll"
     countLowerBuffIndividualitySumSelf = "countLowerBuffIndividualitySumSelf"
     countEqualBuffIndividualitySumPt = "countEqualBuffIndividualitySumPt"
     countEqualBuffIndividualitySumPtAll = "countEqualBuffIndividualitySumPtAll"
     countEqualBuffIndividualitySumOpponent = "countEqualBuffIndividualitySumOpponent"
-    countEqualBuffIndividualitySumOpponentAll = (
-        "countEqualBuffIndividualitySumOpponentAll"
-    )
+    countEqualBuffIndividualitySumOpponentAll = "countEqualBuffIndividualitySumOpponentAll"
     countEqualBuffIndividualitySumSelf = "countEqualBuffIndividualitySumSelf"
     existIndividualityOpponentFront = "existIndividualityOpponentFront"
     existIndividualityOpponentCenter = "existIndividualityOpponentCenter"
@@ -3204,9 +3193,7 @@ class NiceAiCond(str, Enum):
     totalCountHigherIndividualityPt = "totalCountHigherIndividualityPt"
     totalCountHigherIndividualityPtAll = "totalCountHigherIndividualityPtAll"
     totalCountHigherIndividualityOpponent = "totalCountHigherIndividualityOpponent"
-    totalCountHigherIndividualityOpponentAll = (
-        "totalCountHigherIndividualityOpponentAll"
-    )
+    totalCountHigherIndividualityOpponentAll = "totalCountHigherIndividualityOpponentAll"
     totalCountHigherIndividualityAllField = "totalCountHigherIndividualityAllField"
     totalCountLowerIndividualityPt = "totalCountLowerIndividualityPt"
     totalCountLowerIndividualityPtAll = "totalCountLowerIndividualityPtAll"